fix:support redis-py 5 only
