+++ .github/workflows/test.yml
@@ -57,7 +57,7 @@ jobs:
       matrix:
         redis-image: [ "redis:6.2.14", "redis:7.0.15", "redis:7.4.0" ]
         python-version: [ "3.8", "3.9", "3.10", "3.11", "3.12" ]
-        redis-py: [ "4.3.6", "4.6.0", "5.0.8", "5.1.0b7" ]
+        redis-py: [ "4.3.6", "4.6.0", "5.0.9", "5.1.0b7" ]
         include:
           - python-version: "3.12"
             redis-image: "redis/redis-stack:6.2.6-v15"
